+++ src/ralph/models/edx/__init__.py
@@ -24,6 +24,7 @@ from .open_response_assessment.statements import (
     ORASubmitFeedbackOnAssessments,
     ORAUploadFile,
 )
+from .poll.statements import XBlockPollSubmitted, XBlockPollViewResults
 from .peer_instruction.statements import (
     PeerInstructionAccessed,
     PeerInstructionOriginalSubmitted,
@@ -48,6 +49,7 @@ from .problem_interaction.statements import (
     UIProblemShow,
 )
 from .server import Server
+from .survey.statements import XBlockSurveySubmitted, XBlockSurveyViewResults
 from .textbook_interaction.statements import (
     UIBook,
     UITextbookPdfChapterNavigated,